Created first Sphere with a Texture on it
We have now a prototype which has a texture of Earth, some clouds, and has a movable camera attached to it. We accomplished this by using Casey Duncan's Perlin noise for Python library.

Next step is to generate a better Class structure using our UML (currently in progress) and implementing the GUI-Design.

- DM