After a rather long session, we're releasing v1.0.
Yes, we know the last commit has been a while, but all
of this has been done in less time than you'd probably think.
Yes, we know that we should have committed more often to
trace our progress.

Major changes:
- Switched fom Sphinx for Sourcecode documentation to Doxygen
as it's simply easier to use than Sphinx.
- Created a new UML and built it
- README now includes setup and usage
- Requirements will be automatically installed with setup script

So overall, we've met the requirements and should be finished with
the assignment. The only thing left to do for us is finishing up
the protocol and some bug fixing and final toches.